Brexit: A Visual Analytics Approach
In this paper, we look at the 2016 referendum results data. First, will be visually exploring the spatial distributions for different variables from the 2011 U.K. census that we assume to be discriminating, aggregated to local authority level and we will identify the key drivers in England behind the leave vote and compare it with London. Second, will be looking into patterns and relationships using a visual analytics approach. And third, we will use regression analysis to capture how did different ethnics vote in different regions of England using national identity and what characteristics they hold. We will showcase a combination of visual techniques and computational methods across this paper.